refactor(tests): restructure tests/ to mirror src/rfdetr/ layout
- New tests/inference/ (9 files from tests/models/ + 2 from root): predict, from_checkpoint, optimize_for_inference, TRT/keypoint inference; shared _DummyModel/_DummyRFDETR/_BaseFakeRFDETR extracted to tests/inference/helpers.py to remove cross-test-file imports
- New tests/models/heads/ (3 files): detection, keypoint, segmentation head tests
- New tests/models/backbone/test_backbone.py (moved from flat models/)
- New tests/platform/ (test_downloads.py from assets/)
- New tests/training/callbacks/ (4 callback tests)
- tests/export/ gains test_export.py and test_export_for_roboflow.py from models/
- tests/training/ gains test_drop_path.py and test_train_resume.py from models/
- tests/evaluation/ gains test_evaluation.py from root + __init__.py
- tests/models/ gains test_config_keypoints.py and test_namespace.py from root
- _prewarm_dinov2_cache fixture promoted to root tests/conftest.py so inference/ tests also benefit
- Empty dirs (docs/, legacy/, util/) removed

---
