+++ tests/conftest.py
@@ -13,6 +13,25 @@ from rfdetr.datasets.synthetic import DatasetSplitRatios, generate_coco_dataset
 from rfdetr.utilities.reproducibility import seed_all
 
 
+@pytest.fixture(scope="session", autouse=True)
+def _prewarm_dinov2_cache() -> None:
+    """Download DINOv2 backbone weights once per test session.
+
+    HuggingFace hub uses file-level locking internally, so concurrent xdist
+    workers block on each other rather than issuing duplicate network requests.
+    After the first worker finishes, all others read from the local disk cache.
+
+    Examples:
+        This fixture is autouse — no explicit reference needed in tests.
+    """
+    from huggingface_hub import snapshot_download
+
+    snapshot_download(
+        "facebook/dinov2-with-registers-base",
+        ignore_patterns=["*.msgpack", "flax_model*", "tf_model*", "rust_model*"],
+    )
+
+
 @pytest.fixture(autouse=True)
 def reset_random_seeds() -> None:
     """Reset all RNG sources before every test for reproducibility."""
