+++ tests/inference/test_from_checkpoint.py
