+++ tests/inference/test_public_api_keypoints.py
