+++ tests/inference/test_supervision_editable_keypoints.py
