+++ tests/models/heads/__init__.py
@@ -3,4 +3,3 @@
 # Copyright (c) 2025 Roboflow. All Rights Reserved.
 # Licensed under the Apache License, Version 2.0 [see LICENSE for details]
 # ------------------------------------------------------------------------
-"""Shared fixtures for the models test suite."""
