+++ tests/models/test_config_keypoints.py
