+++ tests/models/test_namespace.py
