+++ tests/training/callbacks/test_coco_eval_callback.py
