+++ tests/training/callbacks/test_drop_path_callback.py
