+++ tests/training/callbacks/test_ema_callback.py
